gen_inpaint_dataset/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0

# Diffusers and related
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
safetensors>=0.3.1

# Optional but recommended
xformers>=0.0.16  # For memory-efficient attention
